refactoring hebbian sysnapse and optimizer
