+++ ngclearn/components/optim/sgd.py
@@ -1,4 +1,4 @@
-from ngclearn.utils.optim.opt import Opt
+from ngclearn.components.optim.opt import Opt
 import numpy as np
 from jax import jit, numpy as jnp, random, nn, lax
 from functools import partial
